Added memory logging and reporting, changed to f32

+++ src/graphbuilder.rs
@@ -1,4 +1,4 @@
 pub trait GraphBuilder {
-    fn add_vertex(&mut self, id:  usize, xpos: f64, ypos: f64); 
+    fn add_vertex(&mut self, id:  usize, xpos: f32, ypos: f32); 
 }
